searchFaqの処理を整理
